Moved level related structs and enums to seperate scripts and renamed folder name.
