+++ Assets/Editor/ScriptableObjects/LevelData/Levels.asset
@@ -15,29 +15,237 @@ MonoBehaviour:
   levels:
   - levelID: 3
     song: {fileID: 8300000, guid: 28406c2d4ad2819499dddb87703b8f31, type: 3}
-    level:
+    mappingData:
     - difficulty: 0
       timestamps:
-      - 2
-      - 2.2
-      laneIDs: 0300000003000000
-    - difficulty: 1
-      timestamps:
-      - 2
-      - 2.2
-      laneIDs: 0300000003000000
-    - difficulty: 2
-      timestamps:
-      - 2
-      - 2.2
-      laneIDs: 0300000003000000
+      - 0.46931973
+      - 1.6213152
+      - 2.623991
+      - 3.7119954
+      - 3.8613153
+      - 4.010658
+      - 4.1599774
+      - 4.3306575
+      - 4.458662
+      - 4.607983
+      - 5.247985
+      - 6.2293196
+      - 7.103988
+      - 10.666639
+      - 11.583991
+      - 12.757324
+      - 12.927982
+      - 13.077324
+      - 13.247982
+      - 13.375986
+      - 13.525329
+      - 13.695987
+      - 14.22932
+      - 15.146644
+      - 16.063992
+      - 8.209796
+      - 8.380476
+      - 8.551134
+      - 8.828458
+      - 9.03932
+      - 9.145987
+      - 9.273991
+      - 9.380657
+      - 9.465986
+      - 17.412426
+      - 17.561747
+      - 17.711082
+      - 18.858004
+      - 19.946009
+      - 20.74488
+      - 21.729897
+      - 23.167324
+      - 24.426008
+      - 25.535328
+      - 26.66601
+      - 27.881996
+      - 28.991339
+      - 30.175056
+      - 31.284376
+      - 32.500385
+      - 33.65238
+      - 34.761723
+      - 35.82839
+      - 37.001724
+      - 38.13238
+      - 39.305714
+      - 40.521725
+      - 41.65238
+      - 42.783062
+      - 43.849728
+      - 44.980385
+      - 46.08973
+      - 47.26306
+      - 48.30839
+      - 49.43905
+      - 50.676395
+      - 51.74306
+      - 52.831043
+      - 53.983063
+      - 55.241722
+      - 55.945713
+      - 56.479046
+      - 56.756393
+      - 56.92705
+      - 57.289726
+      - 57.43905
+      - 57.65238
+      - 58.164375
+      - 58.655056
+      - 58.804375
+      - 58.93238
+      - 59.081722
+      - 59.20973
+      - 59.359047
+      - 59.89238
+      - 60.511044
+      - 61.129726
+      - 61.40705
+      - 61.684376
+      - 62.175056
+      - 62.367054
+      - 62.559048
+      - 62.879047
+      - 63.02839
+      - 63.17771
+      - 63.34839
+      - 63.49771
+      - 63.73238
+      - 63.94571
+      - 64.50039
+      - 64.69238
+      - 64.84172
+      - 64.99104
+      - 65.16172
+      - 65.31104
+      - 65.46039
+      - 65.58839
+      - 65.75905
+      - 65.929726
+      - 66.10039
+      - 66.27104
+      - 66.65506
+      - 67.29506
+      - 67.89238
+      - 68.2764
+      - 68.48973
+      - 69.04438
+      - 69.66306
+      - 70.239044
+      - 70.495056
+      - 70.85771
+      - 71.34839
+      - 71.54038
+      - 71.73238
+      - 71.98839
+      - 72.13771
+      - 72.30839
+      - 72.47905
+      - 72.62839
+      - 72.88438
+      - 73.14039
+      - 73.69505
+      - 74.76173
+      - 75.61506
+      - 76.660385
+      - 76.83104
+      - 76.980385
+      - 77.15105
+      - 77.279045
+      - 77.44973
+      - 77.599045
+      - 77.81238
+      - 78.239044
+      - 79.30572
+      - 80.15905
+      - 81.33238
+      - 81.524376
+      - 81.67372
+      - 81.95104
+      - 82.27104
+      - 82.39905
+      - 82.54839
+      - 82.825714
+      - 83.80705
+      - 84.76705
+      - 85.89771
+      - 86.58038
+      - 87.30572
+      - 87.88172
+      - 88.45771
+      - 89.01238
+      - 89.58839
+      - 89.929726
+      - 90.164375
+      - 90.35639
+      - 90.505714
+      - 90.63372
+      - 90.804375
+      - 90.95372
+      - 91.10306
+      - 91.25238
+      - 91.401726
+      - 91.93506
+      - 92.5964
+      - 93.10839
+      - 93.66306
+      - 94.26038
+      - 94.815056
+      - 95.391045
+      - 95.92438
+      - 96.52172
+      - 97.09771
+      - 97.65238
+      - 98.22839
+      - 98.825714
+      - 99.401726
+      - 99.97771
+      - 100.51104
+      - 101.08705
+      - 101.68438
+      - 102.02571
+      - 102.28172
+      - 102.559044
+      - 102.836395
+      - 103.135056
+      - 103.412384
+      - 103.96705
+      - 104.15905
+      - 104.30839
+      - 104.45771
+      - 104.58572
+      - 104.82039
+      - 105.09771
+      - 105.35372
+      - 105.69505
+      - 106.22839
+      - 106.505714
+      - 106.76173
+      - 106.99639
+      - 107.23104
+      - 107.46571
+      - 107.700386
+      - 107.99905
+      - 108.51104
+      - 108.681725
+      - 108.87372
+      - 109.06571
+      - 109.279045
+      laneIDs: 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
   - levelID: 2
     song: {fileID: 8300000, guid: 28406c2d4ad2819499dddb87703b8f31, type: 3}
-    level:
+    mappingData:
     - difficulty: 0
       timestamps:
       - 1
       - 2
-      - 5
-      - 5.5
-      laneIDs: 03000000020000000000000000000000
+      - 2
+      - 3
+      - 3.3
+      - 4
+      laneIDs: 000000000200000001000000030000000300000001000000
