+++ Assets/Scenes/LaneSystem.unity
@@ -1132,14 +1132,7 @@ MonoBehaviour:
   currentLevel:
     levelID: 2
     song: {fileID: 8300000, guid: 28406c2d4ad2819499dddb87703b8f31, type: 3}
-    level:
-    - difficulty: 0
-      timestamps:
-      - 1
-      - 2
-      - 5
-      - 5.5
-      laneIDs: 03000000020000000000000000000000
+    mappingData: []
   difficulty: 0
   okThreshold: 0.75
   alrightThreshold: 0.5
