+++ Assets/Scripts/Framework/Gameplay/Level/Difficulty.cs
@@ -1,0 +1,13 @@
+namespace ProefExamen.Framework.Gameplay.Level
+{
+    /// <summary>
+    /// An enum that is used to express the difficulty of a level.
+    /// </summary>
+    [System.Serializable]
+    public enum Difficulty
+    {
+        Easy = 0,
+        Normal = 1,
+        Hard = 2
+    }
+}
