+++ Assets/Scripts/Framework/Gameplay/Level/MappingData.cs
@@ -1,0 +1,13 @@
+namespace ProefExamen.Framework.Gameplay.Level
+{
+    /// <summary>
+    /// A struct holding mapping data seperatly so that it can be defined for each difficulty.
+    /// </summary>
+    [System.Serializable]
+    public struct MappingData
+    {
+        public Difficulty difficulty;
+        public float[] timestamps;
+        public int[] laneIDs;
+    }
+}
