+++ Assets/Scripts/Framework/Gameplay/Level/MappingData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91b73da927b4d4b41a6969067e5db108
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
